+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [unreleased](https://github.com/spenserblack/github-stats-rs/compare/v0.1.0...master)
 ### Added
 - `is_fork`
+- Language stats for `Repo`
 
 ## 0.1.0 2019/10/08